Create LICENSE
Hi, can I add a license file to the Propy? I am a Bioinformatician student and I used your package  to develop a tool in my master thesis in the area of machine learning. This tool has to be stored in the Galaxy platfrom, but in order to do this all que packages must exist in conda. I took the liberty to add propy to conda (I hope you dont mind) but in order to do this they require a license file, like LGPL. Can you add this lincense file so they can upload the package to conda and I can finish my thesis? 

Thank you very much