DOC: Clearly set GPL v2 as license (#15)
Despite several attempts over more than one year, I was not able to reach the original authors. As I don't know if they want to be connected to propy3, I removed them from the authors list. I will add them back if they want that.

The attribution is given in every single file as well as the git history.